Rename template files to be rendered that are also in gitignore+
The .gitignore file in the main template dir {{cookiecutter.project_name}}
is also actually a template file that is to be copied to the rendered
folder structure. Since the cookiecutter template is also added to a git
repo, the file named .gitignore is treated as the file to ignore files
when processed by git for the cookiecutter template repo.

This was fixed by renaming .gitignore, .env and .vscode to jinja template
variables, which will be rendered to the actual corresponding file names
via defining them in cookiecutter.json